remove a defunct test; minor refactorings
Introduce source-HU to fine-Picking
https://github.com/metasfresh/metasfresh/issues/2298
