+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/picking/pickingCandidateCommands/AddHUToPickingSlot.java
@@ -30,11 +30,11 @@ import lombok.NonNull;
  * 
  * This program is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
  * GNU General Public License for more details.
  * 
  * You should have received a copy of the GNU General Public
- * License along with this program.  If not, see
+ * License along with this program. If not, see
  * <http://www.gnu.org/licenses/gpl-2.0.html>.
  * #L%
  */
@@ -44,10 +44,10 @@ public class AddHUToPickingSlot
 	private final PickingCandidateRepository pickingCandidateRepository;
 
 	public AddHUToPickingSlot(@NonNull final PickingCandidateRepository pickingCandidateRepository)
-	{this.pickingCandidateRepository = pickingCandidateRepository;
-		
+	{
+		this.pickingCandidateRepository = pickingCandidateRepository;
 	}
-	
+
 	public void perform(final int huId, final int pickingSlotId, final int shipmentScheduleId)
 	{
 		// Get HU's single product storage
@@ -81,7 +81,7 @@ public class AddHUToPickingSlot
 		}
 
 		final I_M_Picking_Candidate pickingCandidate = pickingCandidateRepository.getCreateCandidate(huId, pickingSlotId, shipmentScheduleId);
-		
+
 		pickingCandidate.setQtyPicked(qty);
 		pickingCandidate.setC_UOM(uom);
 		save(pickingCandidate);
