+++ de.metas.handlingunits.base/src/main/java/de/metas/handlingunits/picking/pickingCandidateCommands/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * Contains business logic that is invoked by {@link de.metas.handlingunits.picking.PickingCandidateCommand}
+ * 
+ * @author metas-dev <dev@metasfresh.com>
+ *
+ */
+package de.metas.handlingunits.picking.pickingCandidateCommands;
