+++ de.metas.handlingunits.base/src/test/java/de/metas/handlingunits/picking/impl/RetrieveAvailableHUsToPickFiltersTests.java
@@ -19,6 +19,7 @@ import de.metas.handlingunits.model.I_M_Source_HU;
 import de.metas.handlingunits.model.X_M_HU;
 import de.metas.handlingunits.model.X_M_HU_Item;
 import de.metas.handlingunits.picking.impl.HUPickingSlotBL;
+import de.metas.handlingunits.picking.impl.HUPickingSlotBLs.RetrieveAvailableHUsToPickFilters;
 
 /*
  * #%L
